Deploy ngrok

+++ src/Books/books.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Books } from './books.entity';
@@ -8,9 +12,7 @@ import { title } from 'process';
 
 @Injectable()
 export class BooksService {
-  constructor(
-    private booksRepository: BooksRepository,
-  ) {}
+  constructor(private booksRepository: BooksRepository) {}
 
   async findAll(): Promise<Books[]> {
     return this.booksRepository.findAll();
@@ -29,17 +31,26 @@ export class BooksService {
     if (!existingBook) {
       throw new NotFoundException(`Book with ID ${id} not found`);
     }
+    if (existingBook.author !== book.author) {
+      throw new ForbiddenException(
+        `You do not have permission to delete this book`,
+      );
+    }
     await this.booksRepository.update(id, book);
     return this.booksRepository.findOne(title);
   }
 
-  async delete(id: number): Promise<BooksDto | undefined> {
+  async delete(title: string, author: string): Promise<BooksDto | undefined> {
     const existingBook = await this.booksRepository.findOne(title);
     if (!existingBook) {
-      throw new NotFoundException(`Book with ID ${id} not found`);
+      throw new NotFoundException(`Book with ID ${title} not found`);
     }
-
+    if (existingBook.author !== author) {
+      throw new ForbiddenException(
+        `You do not have permission to delete this book`,
+      );
+    }
     await this.booksRepository.delete(title);
-    return this.booksRepository.findOne(title);
+    return undefined;
   }
 }
